pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_prom_vid"
version = "0.1.0"
description = "Automated product promotional video generator"
authors = [ { name="Your Name" } ]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["video","automation","scraping","tts","llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "requests",
  "beautifulsoup4",
  "lxml",
  "pydantic>=2.0.0",
  "python-dotenv",
  "openai>=1.0.0",
  "moviepy",
  "Pillow",
  "edge-tts; platform_system == 'Windows'",
  "tqdm"
]

[project.optional-dependencies]
dev = ["pytest","ruff","mypy"]

[project.scripts]
auto-prom-vid = "pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
